[issue #8] proper deselect if ws is root ws
The node key builder fails in case of a root ws because the key is
initialised with the id of the ws which is often an integer. As a root
ws does not have a parent the key remains the id of the initial ws which
in turn is treated as an integer thus not having a split operation.
Simply converting the id to a string is sufficient to solve this.